+++ tests/test_devices.py
@@ -1827,3 +1827,8 @@ def test_wall_mounted_keypad(fake_home: Home):
     with no_ssl_verification():
         d = fake_home.search_device_by_id("3014F7110000000000000WKP")
         assert isinstance(d, WallMountedKeyPad)
+
+def test_watering_actuator(fake_home: Home):
+    with no_ssl_verification():
+        d = fake_home.search_device_by_id("3014F71100000000000SHWSM")
+        assert isinstance(d, WateringActuator)